fibonacci 1

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
 int main(void)
 {
 	int i;
-	long int fibonacci[50];
+	int fibonacci[50];
 
 	fibonacci[0] = 1;
 	fibonacci[1] = 2;
